+++ cli/cat.c
@@ -49,9 +49,10 @@ int tec_cli_cat(tec_argvec_t *argvec, tec_ctx_t *ctx)
             opt_quiet = true;
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }
     i = optind;