+++ cli/cd.c
@@ -39,9 +39,10 @@ int tec_cli_cd(tec_argvec_t *argvec, tec_ctx_t *ctx)
             opt_cd_toggle = false;
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }