+++ cli/desk.c
@@ -76,9 +76,10 @@ static int _desk_add(tec_argvec_t *argvec, tec_ctx_t *ctx)
             opt_cd_toggle = false;
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }
     i = optind;
@@ -180,9 +181,10 @@ static int _desk_rm(tec_argvec_t *argvec, tec_ctx_t *ctx)
             opt_ask_once = true;
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }
     i = optind;
@@ -252,9 +254,10 @@ static int _desk_ls(tec_argvec_t *argvec, tec_ctx_t *ctx)
             opt_help = true;
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }
 
@@ -326,9 +329,10 @@ static int _desk_set(tec_argvec_t *argvec, tec_ctx_t *ctx)
             ctx->units = tec_unit_add(ctx->units, "desc", optarg);
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }
 
@@ -386,9 +390,10 @@ static int _desk_cat(tec_argvec_t *argvec, tec_ctx_t *ctx)
             opt_quiet = true;
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }
     i = optind;
@@ -454,9 +459,10 @@ static int _desk_cd(tec_argvec_t *argvec, tec_ctx_t *ctx)
             opt_cd_toggle = false;
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }