+++ cli/env.c
@@ -78,9 +78,10 @@ static int _env_add(tec_argvec_t *argvec, tec_ctx_t *ctx)
             switch_env = false;
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }
 
@@ -214,9 +215,10 @@ static int _env_rm(tec_argvec_t *argvec, tec_ctx_t *ctx)
             opt_ask_once = true;
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }
     i = optind;
@@ -284,9 +286,10 @@ static int _env_ls(tec_argvec_t *argvec, tec_ctx_t *ctx)
             opt_help = true;
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }
 
@@ -335,7 +338,7 @@ static int _env_rename(tec_argvec_t *argvec, tec_ctx_t *ctx)
             opt_quiet = true;
             break;
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }
 
@@ -403,9 +406,10 @@ static int _env_set(tec_argvec_t *argvec, tec_ctx_t *ctx)
             ctx->units = tec_unit_add(ctx->units, "desc", optarg);
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }
 
@@ -461,9 +465,10 @@ static int _env_cat(tec_argvec_t *argvec, tec_ctx_t *ctx)
             quiet = true;
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }
 
@@ -524,9 +529,10 @@ static int _env_cd(tec_argvec_t *argvec, tec_ctx_t *ctx)
             opt_cd_toggle = false;
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }