+++ cli/help.c
@@ -723,9 +723,10 @@ int tec_cli_help(tec_argvec_t *argvec, tec_ctx_t *ctx)
             helpctx.desc_long = false;
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         };
     }