+++ cli/mv.c
@@ -295,9 +295,10 @@ int tec_cli_mv(tec_argvec_t *argvec, tec_ctx_t *ctx)
             return elog(1, "option `-t' under development");
             break;
         case ':':
-            return elog(1, "option `-%c' requires an argument", optopt);
+            return elog(EXIT_FAILURE, "option '-%c' requires an argument",
+                        optopt);
         default:
-            return elog(1, "invalid option `-%c'", optopt);
+            return elog(EXIT_FAILURE, "invalid option '-%c'", optopt);
         }
     }